refactor: fix critical bugs, remove dead code, extract hooks and shared components
Fix Section forwardRef bug that broke all Toolbox scroll/focus navigation.
Fix missing requestAnimationFrame cleanup in focus effects. Remove erroneous
no-print class from Evidence article. Extract useMobileMenu, useNavigationFocus,
and useDownloadHandlers from App.jsx (693 → ~250 lines). Create reusable
SectionHeader and AsideCard components replacing ~15 duplicated patterns across
all templates. Remove ~590 lines of unused CSS (home-*, print-shell, etc.) and
15 unused data exports from evidenceContent.js. Net reduction: ~1600 lines.
