+++ src/index.css
@@ -156,19 +156,6 @@
     min-width: 0;
   }
 
-  .editorial-shell {
-    position: relative;
-  }
-
-  .editorial-shell::after {
-    content: '';
-    position: absolute;
-    inset: 0;
-    pointer-events: none;
-    border-radius: var(--radius-xl);
-    box-shadow: inset 0 1px 0 color-mix(in srgb, white 80%, transparent);
-  }
-
   @media (min-width: 640px) {
     .container {
       padding-left: 1.5rem;
